Expand information on the "Tasks" plugin
I added a link to the docs, a summary, and screenshots to the Tasks
plugin's page to make it easier for people to see what it is about when
they browse the community hub's plugins.

I tried not to add too much information so that this page won't need
updating with future releases of Tasks.