Ensure that block-links recognised after comments deleted
It turns out that having any white-space after a block name (such as ^github)
breaks the block link, and treats `^github ` as ordinary text.

To make this less likely to happen, I've removed the spaces before any closing
comments on links with block names.

(There are no cases of this in current pages, which is either because
those editing pages know the importance of removing spaces after
block names - or perhaps because not many template pages have beeen
edited. Or a combination of both?)