"Unlisted plugins" and "Others" now hidden with <!-- -->
These sections each have their own pairs of comments
now, so it's easier to activate them individually.

As well as making it more obvious in Obsidian that these
sections are commented out, it also makes
it easier to see the matching pairs of comments,
and so easier to see which lines to delete, to activate
a section.