Remove a test that saved the community theme download counts
This test was always going to be unstable over time, as new themes are added.

However, it was really useful when I was first adding type hints,
as it showed how the community themes downloads list was represented in Python.

I've added some comments in get_theme_downloads() to retain the
knowledge about how the theme download data is structured.