Theme class instead of inheritance
For easier code maintenance and code comprehension, we want to encapsulate theme-related code inside a Theme class.

A step on the way is to hide Theme's data storage. This change makes Theme
store a dict, instead of inheriting one. It still provides a dict interface, so more encapsulation will be needed later.

See #151

Co-Authored-By: kometenstaub <83140328+kometenstaub@users.noreply.github.com>