Added description and changed link
I’ve added much more explanation on how to install and use the shortcut. In addition, I’ve changed the link. This goes to the same place as the previous link, but allows me to easily redirect it if I change where the shortcut is hosted; Routinehub, where it’s hosted currently, has been extremely slow lately.