Silence error output on console for correct test of invalid theme.
The test is working fine, it is just outputting what looks like an error
to the console when it runs. The test itself is testing an invalid case,
so that is expected behavior, but we don't need to see the expected
error test case console output.

I wrote a little wrapper silence_stdout to suppress the console output
just for the bit of the test that produces the confusing looking output.

All the test assertion related output is still printed to the console,
and if someone needed to see more details when debugging this test
commenting out the `with silence_stdout():` line (and de-indenting the
contents of that block) should be easy.