Add code to automatically alphabetize plugin lists by alias.
If there is no alias the page name is used for sorting.
This is meant to be re-runnable and handle links with or without
discussions.

It is written to work pretty generically on the first markdown list
within a heading on a page, so we could re-apply this concept to other
sets of pages and headings.