Fix add_footers script.
the first line of ensure_footer_in_file was trying to calculate
a relative path, based on an already relative path. which introduced the
../../github/scripts into the paths which we do not want in the
templates.

This also alters the call to regex_replace_in_file to pass the
absolute_path to the file, not the relative path, which fixes the
read_file error.