Regenerate Lumines theme note for new author
This is being done to update the theme author from Avesend to danielkhmara

It replaces the edit in #784, which was giving a merge conflict due to changes in the number of download counts.

As a side effect, the script also generated details of the available Style Settings options.

(I couldn't safely push the conflict resolution to the author's branch in the PR, because the PR was created from the author's 'main' branch.)